requirements.txt
beautifulsoup4
lxml
requests
python-Levenshtein
